refactor(search): merged SearchInput and MusicSearchInput by giving them props
